apply review feedback: error-path hardening and always-on tests
- MacFileMonitor: fall through to per-path processing when the
  drop-recovery rescan fails, instead of silently discarding the batch
- ClientEventDispatcher: null-guard the files_changed payload and
  dispatch each element independently so one failure doesn't drop the
  rest of the batch
- VcsState: extract the batched merge into mergeFileChanges() and pin
  its behavior with a new VcsStateTests GWTTestCase
- e2e: drop the PW_PERF gate; the spec is now a functional test at
  tests/panes/git/vcs_flood.test.ts that always runs -- timing is
  logged but not asserted, and batched files_changed delivery, the
  post-burst probe, and the .gitignore full-refresh path are asserted
- reword the O_CLOEXEC comment (closes on exec, not fork)
